thau-cli/requirements.txt
click>=8.0.0
requests>=2.31.0
websockets>=11.0.0
rich>=13.0.0
prompt_toolkit>=3.0.0
pyyaml>=6.0.0
aiohttp>=3.8.0
